--- .gitattributes
@@ -1,1 +1,0 @@
-src/*/runtimes/**/* filter=lfs diff=lfs merge=lfs -text
