--- .github/workflows/embed-binaries.yaml
@@ -1,20 +1,0 @@
-name: Embed Binaries
-
-on:
-  schedule:
-    - cron: "59 23 * * *"
-  workflow_dispatch:
-
-permissions:
-  issues: write
-
-jobs:
-  embed-binaries:
-    uses: devantler-tech/workflows/.github/workflows/dotnet-embed-binaries.yaml@main
-    secrets: inherit
-    with:
-      BINARY_NAME: cilium
-      DOTNET_PROJECT_NAME: Devantler.CiliumCLI
-      SCRIPT_PATH: scripts/get-cilium-binary.sh
-      VERSION_ARGS: version --client | awk '{print $2}' | head -1
-
