--- .github/workflows/readme-tree.yaml
@@ -1,15 +1,0 @@
-name: Readme Tree
-
-on:
-  workflow_dispatch:
-  push:
-    branches: [main]
-
-permissions:
-  contents: write
-  pull-requests: write
-
-jobs:
-  readme-tree:
-    uses: devantler-tech/workflows/.github/workflows/readme-tree.yaml@main
-    secrets: inherit
