--- src/Devantler.CiliumCLI/runtimes/linux-arm64/native/cilium-linux-arm64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:72af957e0c320756e6d2e3cffa0d8deb62ed4d19fd5ca9f5860ac92ff2305b57
-size 148242616
