--- src/Devantler.CiliumCLI/runtimes/linux-x64/native/cilium-linux-x64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:aeee345f51fdb870ad1412e234f55a15a68f564294bb91b5a6daa7d6b455e2c6
-size 152776888
