--- src/Devantler.CiliumCLI/runtimes/osx-arm64/native/cilium-osx-arm64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b70b7e2210f5d076692b3ece2dd35cb8858eb4c5f4ab4c0f7a4d9bdb2ced0098
-size 152064082
