--- src/Devantler.CiliumCLI/runtimes/osx-x64/native/cilium-osx-x64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2465582a7900602460f672b1da37b77f9a36d7ae71333a802570091cfea86479
-size 155489376
